+++ src/main/java/org/cactoos/text/HexOf.java
@@ -35,7 +35,7 @@ public final class HexOf extends TextEnvelope {
                     final char[] hex = new char[bts.length * 2];
                     int chr = -1;
                     for (final byte currentbyte: bts) {
-                        final int value = 0xff & (int) currentbyte;
+                        final int value = 0xff & currentbyte;
                         hex[++chr] = HexOf.HEX_CHARS[value >>> 4];
                         hex[++chr] = HexOf.HEX_CHARS[value & 0x0f];
                     }
